refatorando3
