+++ include/CMakeLists.txt
@@ -2,4 +2,8 @@ target_sources(${PROJECT_NAME} PRIVATE
     Color.hpp
     Ray.hpp
     Vector3d.hpp
+    Camera.hpp
+    Lights.hpp
+    Objects.hpp
+    Scene.hpp
 )
