+++ include/Color.hpp
@@ -3,36 +3,39 @@
 
 #include <iostream>
 #include <cstdint>
-class Color;
 
-class IntensityColor
-{
-public:
-    float r;
-    float g;
-    float b;
+namespace atividades_cg_1::color {
+    class Color;
+    class IntensityColor
+    {
+    public:
+        float r;
+        float g;
+        float b;
 
-    IntensityColor();
-    IntensityColor(float r, float g, float b);
+        IntensityColor();
+        IntensityColor(float r, float g, float b);
 
-    IntensityColor arroba_multiply(IntensityColor intensity);
-    IntensityColor multiply(float value);
-    IntensityColor sum(IntensityColor other);
-    IntensityColor minus(IntensityColor other);
-    Color to_color();
-};
-class Color
-{
-public:
-    uint8_t r;
-    uint8_t g;
-    uint8_t b;
+        IntensityColor arroba_multiply(IntensityColor intensity);
+        IntensityColor multiply(float value);
+        IntensityColor sum(IntensityColor other);
+        IntensityColor minus(IntensityColor other);
+        Color to_color();
+    };
 
-    Color();
-    Color(uint8_t red, uint8_t green, uint8_t blue);
+    class Color
+    {
+    public:
+        uint8_t r;
+        uint8_t g;
+        uint8_t b;
 
-    IntensityColor to_color_normalized() const;
-    Color multiply(IntensityColor intensity) const;
-};
+        Color() {}
+        Color(uint8_t red, uint8_t green, uint8_t blue): r(red), g(green), b(blue) {}
 
-#endif;
+        IntensityColor to_color_normalized() const;
+        Color multiply(IntensityColor intensity) const;
+    };
+
+}
+#endif
