+++ src/Ray.cpp
@@ -3,7 +3,7 @@
  #include "Vector3d.hpp"
  #include "Ray.hpp"
 
-using namespace algebra;
+using namespace atividades_cg_1::algebra;
 
 
 float Ray::size()
@@ -16,12 +16,4 @@ Vector3d Ray::get_dr()
 {
     return (p2.minus(p1)).divide(this->size());
 }
-
-
-// std::ostream &operator<<(std::ostream &os, Ray const &r)
-// {
-//     return os << "Ray(P1: " << r.p1 << " | P2: " << r.p2 << ")";
-// }
-
-
