+++ src/main.cpp
@@ -6,321 +6,16 @@
 #include "Color.hpp"
 #include "Vector3d.hpp"
 #include "Ray.hpp"
+#include "Objects.hpp"
+#include "Camera.hpp"
+#include "Scene.hpp"
 
 using namespace std;
 
-using namespace algebra;
-
-class SourceOfLight
-{
-public:
-    IntensityColor intensity;
-    Vector3d center;
-    SourceOfLight() {}
-    SourceOfLight(IntensityColor intensity, Vector3d center)
-    {
-        this->intensity = intensity;
-        this->center = center;
-    }
-    SourceOfLight(Color color, Vector3d center)
-    {
-        this->intensity = color.to_color_normalized();
-        this->center = center;
-    }
-};
-
-
-class Intersection
-{
-public:
-    float time;
-    bool is_valid;
-
-    Intersection() {}
-    Intersection(float t, bool valid)
-    {
-        this->time = t;
-        this->is_valid = valid;
-    }
-};
-
-class Object
-{
-public:
-    virtual ~Object() {}
-    // bool intersected_by(Ray ray) {}
-    Color color;
-    IntensityColor difuse_reflectivity;   // K_d
-    IntensityColor specular_reflectivity; // K_e
-    IntensityColor environment_reflectivity; // K_a
-    float shininess;
-    virtual Intersection get_intersection(Ray ray) { return Intersection(0.0, false); }
-    virtual Vector3d get_normal_vector(Vector3d intersection_point) { return Vector3d(20,20,20);};
-    virtual Vector3d get_light_vector(Vector3d intersection_point, SourceOfLight source_of_light) {return Vector3d(20,20,20); }
-
-    IntensityColor get_difuse_contribution(Vector3d intersection_point, SourceOfLight source_of_light)
-    {
-
-        Vector3d l = this->get_light_vector(intersection_point, source_of_light);
-        Vector3d n = this->get_normal_vector(intersection_point);
-        float scalar_product_l_n = l.scalar_product(n);
-
-        // if scalar_product is negative, the angle is greater than 180 degrees and we can't see the object. The color in result will be black.
-        if (scalar_product_l_n < 0)
-        {
-            scalar_product_l_n = 0;
-        }
-        float f_difuse = scalar_product_l_n;
-        IntensityColor contribution = source_of_light.intensity.arroba_multiply(this->difuse_reflectivity).multiply(f_difuse);
-
-        return contribution;
-    }
-
-    IntensityColor get_specular_contribution(Vector3d intersection_point, Vector3d eye_point, SourceOfLight source_of_light)
-    {
-        Vector3d l = this->get_light_vector(intersection_point, source_of_light);
-        Vector3d n = this->get_normal_vector(intersection_point);
-
-        Vector3d r = n.multiply(2).minus(l).get_vector_normalized();
-        Vector3d v = eye_point.minus(intersection_point).get_vector_normalized();
-
-        float f_specular = std::pow(r.scalar_product(v), this->shininess);
-
-        IntensityColor contribution = source_of_light.intensity.arroba_multiply(this->specular_reflectivity).multiply(f_specular);
-        return contribution;
-    }
-};
-
-class Sphere : public Object
-{
-public:
-    Vector3d center;
-    float radius;
-
-    Sphere(Vector3d center, float radius, Color color, IntensityColor difuse_reflectivity, IntensityColor specular_reflectivity, IntensityColor environment_reflectivity, float shininess)
-    {
-        this->center = center;
-        this->radius = radius;
-        this->color = color;
-        this->difuse_reflectivity = difuse_reflectivity;
-        this->specular_reflectivity = specular_reflectivity;
-        this->environment_reflectivity = environment_reflectivity;
-        this->shininess = shininess;
-    }
-
-    // n unitary vector (normal vector).
-    Vector3d get_normal_vector(Vector3d intersection_point) override
-    {
-        return (intersection_point.minus(this->center)).divide(this->radius);
-    }
-
-    // l unitary vector in direction to light.
-    Vector3d get_light_vector(Vector3d intersection_point, SourceOfLight source) override
-    {
-        return (source.center.minus(intersection_point)).get_vector_normalized();
-    }
-
-    Intersection get_intersection(Ray ray) override
-    {
-
-        Vector3d initial_point = ray.p1;
-        Vector3d dr = ray.get_dr();
-
-        Vector3d w = initial_point.minus(this->center);
-
-        // We want to check if ||Pint - C||² = R²
-        // (Pin + tint*dr - C).(Pin + tint*dr - C) - R² = 0
-        // (w + tint*dr).(w + tint*dr) - R² = 0
-        // w² + 2w*tint*dr + tint²*dr² - R² = 0
-        // dr²*tint² + 2w*dr*tint + w² -     R² = 0
-        float a = dr.scalar_product(dr);
-        float b = (w.multiply(2)).scalar_product(dr);
-        float c = w.scalar_product(w) - std::pow(this->radius, 2);
-
-        float delta = std::pow(b, 2) - (4 * a * c);
-        float t1 = (-b + std::sqrt(delta)) / (2 * a);
-        float t2 = (-b - std::sqrt(delta)) / (2 * a);
-
-        if (delta < 0)
-        {
-            return Intersection(0.0, false);
-        }
-
-        Vector3d intersection_point;
-        if (t1 < t2)
-            return Intersection(t1, true);
-        return Intersection(t2, true);
-    }
-};
-
-class Plan: public Object {
-public:
-    Vector3d known_point;
-    Vector3d normal;
-
-    Plan(Vector3d known_point, Vector3d normal, IntensityColor difuse_reflectivity, IntensityColor specular_reflectivity, IntensityColor environment_reflectivity, float shininess, Color color) {
-        this->known_point = known_point;
-        this->normal = normal;
-        this->difuse_reflectivity = difuse_reflectivity;
-        this->specular_reflectivity = specular_reflectivity;
-        this->environment_reflectivity = environment_reflectivity;
-        this->shininess = shininess;
-        this->color = color;
-    }
-
-    Vector3d get_normal_vector(Vector3d intersection_point) override {
-        return this->normal.get_vector_normalized();
-    }
-    Vector3d get_light_vector(Vector3d intersection_point, SourceOfLight source_of_light) override {
-        return source_of_light.center.minus(intersection_point).get_vector_normalized();
-    }
-
-    Intersection get_intersection(Ray ray) override {
-        Vector3d w = ray.p1.minus(this->known_point);
-//        Vector3d w = this->known_point.minus(ray.p2);
-        Vector3d dr = ray.get_dr();
-
-        float t_int = -(this->normal.scalar_product(w))/(this->normal.scalar_product(dr));
-
-        return Intersection(t_int, t_int > 0);
-    }
-};
-
-class Scene
-{
-public:
-    std::vector<Object *> objects;
-    Color background_color;
-    SourceOfLight source_of_light;
-    IntensityColor environment_light;
-    Vector3d eye;
-
-    Scene(Color bg_color, SourceOfLight source, IntensityColor environment_light, Vector3d eye)
-    {
-        this->background_color = bg_color;
-        this->source_of_light = source;
-        this->environment_light = environment_light;
-        this->eye = eye;
-    }
-
-    void push_object(Object *obj)
-    {
-        objects.push_back(obj);
-    }
-
-    Color get_color_to_draw(Ray ray)
-    {
-        float min_time_intersection = 10000000.0;
-        int min_index = -1;
-        for (int i = 0; i < objects.size(); i++)
-        {
-            Intersection intersection = objects[i]->get_intersection(ray);
-            if (intersection.is_valid)
-            {
-                if (intersection.time < min_time_intersection)
-                {
-                    min_time_intersection = intersection.time;
-                    min_index = i;
-                }
-            }
-        }
-
-        if (min_index == -1)
-            return this->background_color;
-
-        Object *obj = objects[min_index];
-        Color color = obj->color;
-
-        // Pin + t*dr
-        Vector3d intersection_point = ray.p1.sum(ray.get_dr().multiply(min_time_intersection));
-
-        // Check if the object is seen by pontual light
-        Ray ray_light(this->source_of_light.center, intersection_point);
-
-        float min_t2 = 10000000.0;
-        Object *min_obj = objects[0];
-
-        for (auto & object : objects)
-        {
-            Intersection intersection2 = object->get_intersection(ray_light);
-            if (intersection2.is_valid)
-            {
-                if (intersection2.time < min_t2)
-                {
-                    min_t2 = intersection2.time;
-                    min_obj = object;
-                }
-            }
-        }
-
-        Vector3d intersection_point2 = ray_light.p1.sum(ray_light.get_dr().multiply(min_t2));
-
-        // If intersection times are not equal, there is something interrupting light to get into that point, so we discard difuse and specular contributions.
-        if (min_obj != obj) {
-            IntensityColor environment_contrib = this->environment_light.arroba_multiply(obj->environment_reflectivity);
-
-            return color.multiply(environment_contrib);
-        }
-
-        IntensityColor difuse_contrib = obj->get_difuse_contribution(intersection_point, source_of_light);
-        IntensityColor specular_contrib = obj->get_specular_contribution(intersection_point, this->eye, source_of_light);
-        IntensityColor environment_contrib = this->environment_light.arroba_multiply(obj->environment_reflectivity);
-
-        IntensityColor result = environment_contrib.sum(difuse_contrib).sum(specular_contrib);
-
-
-        return color.multiply(result);
-    }
-
-    void dealloc_objects()
-    {
-        for (auto &obj : objects)
-        {
-            delete obj;
-        }
-    }
-};
-
-class Window
-{
-public:
-    Vector3d center;
-    float width;
-    float height;
-
-    int cols;
-    int rows;
-
-    // dx and dy of rectangles
-    float dx;
-    float dy;
-
-    std::vector<std::vector<Color>> windows_colors;
-    bool should_update = true;
-
-    Window() {}
-    Window(float width, float height, int cols, int rows, float x, float y, float z)
-    {
-        this->width = width;
-        this->height = height;
-        this->cols = cols;
-        this->rows = rows;
-        this->center = Vector3d(x, y, z);
-
-        this->dx = width / (float)cols;
-        this->dy = height / (float)rows;
-
-        for (int i = 0; i < rows; i++)
-        {
-            std::vector<Color> row;
-            for (int j = 0; j < cols; j++)
-            {
-                row.push_back(Color(100, 100, 100));
-            }
-            windows_colors.push_back(row);
-        }
-    }
-};
+using namespace atividades_cg_1::algebra;
+using namespace atividades_cg_1::objects;
+using namespace atividades_cg_1::camera;
+using namespace atividades_cg_1::scene;
 
 
 int render_picture(int n_rows, int n_cols, int sdl_width, int sdl_height, float window_width, float window_height)
